Sync with upstream (#1)
* Handle username changes correctly

If a user reauthenticated after changing their GitHub username, this
code would create a second user record for them. The current_user method
returns the oldest matching user record, so the install script would
contain the wrong username and wouldn't work properly.

With this change we'll update the existing user record instead.

* Restrict access by GitHub organization

This allows you to restrict access to members of a GitHub organisation,
similar to the existing feature restricting access to members of a team.

* Allow to be run on macOS Sierra

This adds support so that you can run Boxen on macOS Sierra (10.12).